mds-interface-agent/pyproject.toml
[project]
name = "mds-interface-agent"
version = "1.0.0"
description = "AI agent for MDS 9710 Fibre Channel interface triage"
requires-python = ">=3.11"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["app*", "tools*"]
